Notebook for collab
Reorder everythging for collab
